+++ server/pkg/api/spot/v1/get_players_states_test.go
@@ -159,5 +159,5 @@ func TestGetPlayerPosition(t *testing.T) {
 	assert.NotNil(t, mockServer.LastPlayersPositions)
 	assert.Equal(t, float64(10), mockServer.LastPlayersPositions.PlayerState.Position.Latitude)
 	assert.Equal(t, float64(20), mockServer.LastPlayersPositions.PlayerState.Position.Longitude)
-	assert.Equal(t, int32(88), mockServer.LastPlayersPositions.PlayerState.Health)
+	assert.Equal(t, float32(88), mockServer.LastPlayersPositions.PlayerState.Health)
 }
