+++ server/pkg/zone/zone_test.go
@@ -20,6 +20,6 @@ func TestNewZone(t *testing.T) {
 	z := NewZone(s2.LatLng{Lat: 0, Lng: 0}, 200, 10)
 
 	assert.Equal(t, s2.LatLng{Lat: 0, Lng: 0}, z.Position)
-	assert.Equal(t, uint32(200), z.Radius)
+	assert.Equal(t, float32(200), z.Radius)
 	assert.Equal(t, float32(0.25), z.Damage)
 }
